+++ src/spectrogram/internal/globalspectrogramconfiguration.h
@@ -15,13 +15,13 @@ public:
 
     void init();
 
-    int minFreq() const override;
-    void setMinFreq(int value) override;
-    muse::async::Channel<int> minFreqChanged() const override;
+    double minFreq() const override;
+    void setMinFreq(double value) override;
+    muse::async::Channel<double> minFreqChanged() const override;
 
-    int maxFreq() const override;
-    void setMaxFreq(int value) override;
-    muse::async::Channel<int> maxFreqChanged() const override;
+    double maxFreq() const override;
+    void setMaxFreq(double value) override;
+    muse::async::Channel<double> maxFreqChanged() const override;
 
     bool spectralSelectionEnabled() const override;
     void setSpectralSelectionEnabled(bool value) override;
@@ -66,8 +66,8 @@ public:
     muse::async::Notification someSettingChanged() const override;
 
 private:
-    muse::async::Channel<int> m_minFreqChanged;
-    muse::async::Channel<int> m_maxFreqChanged;
+    muse::async::Channel<double> m_minFreqChanged;
+    muse::async::Channel<double> m_maxFreqChanged;
     muse::async::Channel<bool> m_spectralSelectionEnabledChanged;
     muse::async::Channel<SpectrogramColorScheme> m_colorSchemeChanged;
     muse::async::Channel<int> m_colorGainDbChanged;
